% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{autoc}
\alias{autoc}
\title{Run-off triangles of Personal Auto and Commercial Auto insurance, obtained from the
   ChainLadder package}
\description{
A list of three matrices, paid Personal Auto, incurred Personal Auto and paid
Commercial Auto respectively.
}
\references{
Gesmann M., Murphy D., Zhang W., ChainLadder: Statistical methods and models for the calculation
 of outstanding claims reserves in general insurance.
}
\keyword{datasets}

