% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mortgage}
\alias{mortgage}
\title{Dataset from Thomas Mack's 1999 paper}
\description{
Dataset from Thomas Mack's 1999 paper as in the ChainLadder package
}
\references{
Mack T., The Standard Error of Chain Ladder Reserve Estimates Recursive Calculation
   and Inclusion of a tail factor, ASTIN BULLETIN, vol 29, No. 2, 1999, pp 361 - 366.
}
\keyword{datasets}

